Make reporting less biased by rerunning winner in random situation
Before, we'd show the winning trace of the winning genome,
which would probably be a lucky run. Now we take the winner but
make them rerun in a random situation.

Conversely, now the behavior is quite weird, where the average score
of a genome might go up, but the trace we return to visualize makes it
look worse.